backend/requirements.txt
django==3.1.2
djangorestframework==3.11.2 
markdown==3.3 
django-filter==2.4.0
django_cors_headers==3.5.0
psycopg2-binary>=2.8
black==22.3.0
Pillow==7.1.0
isort[requirements]==4.3.21
pylint==2.7.4
watchdog[watchmedo]==0.10.3
django-autoslug==1.9.8
django-model-utils==4.1.1
sentry-sdk==1.4.3
python-decouple==3.5
gunicorn==20.1.0
funcy==1.14
ipdb==0.13.9
django-phonenumber-field==4.0.0
phonenumbers==8.12.7
